requirements.txt
Flask
flask-cors-3.0.8
pillow-7.1.1
